faster image warping & creation of synthetic task

+++ src/vision/bilinear.h
@@ -10,11 +10,7 @@ namespace nano
         ///
         /// NB: the output matrix is considered to be already resized to the desired size.
         ///
-        template
-        <
-                typename tmatrixi,
-                typename tmatrixo
-        >
+        template <typename tmatrixi, typename tmatrixo>
         void bilinear(const tmatrixi& srcplane, tmatrixo&& dstplane)
         {
                 assert(srcplane.rows() > 0);
@@ -25,20 +21,20 @@ namespace nano
                 const int orows = static_cast<int>(dstplane.rows());
                 const int ocols = static_cast<int>(dstplane.cols());
 
-                const double scale_rows = double(irows) / double(orows);
-                const double scale_cols = double(icols) / double(ocols);
+                const float scale_rows = float(irows) / float(orows);
+                const float scale_cols = float(icols) / float(ocols);
 
                 for (int _or = 0; _or < orows; ++ _or)
                 {
-                        const double isr = scale_rows * _or;
+                        const float isr = scale_rows * _or;
                         const int ir0 = static_cast<int>(isr), ir1 = std::min(ir0 + 1, irows - 1);
-                        const double wr1 = isr - ir0, wr0 = 1.0 - wr1;
+                        const float wr1 = isr - ir0, wr0 = 1 - wr1;
 
                         for (int _oc = 0; _oc < ocols; ++ _oc)
                         {
-                                const double isc = scale_cols * _oc;
+                                const float isc = scale_cols * _oc;
                                 const int ic0 = static_cast<int>(isc), ic1 = std::min(ic0 + 1, icols - 1);
-                                const double wc1 = isc - ic0, wc0 = 1.0 - wc1;
+                                const float wc1 = isc - ic0, wc0 = 1 - wc1;
 
                                 dstplane(_or, _oc) = static_cast<typename tmatrixo::Scalar>(
                                         wr0 * wc0 * srcplane(ir0, ic0) +
@@ -49,4 +45,3 @@ namespace nano
                 }
         }
 }
-
